1. viewer화면 장비보기  -> 각 poi category클릭하면 해당 건물의 poi toggle 처리 2. viewer -> 건물 클릭 -> /map으로 이동(임시) 3. map화면에서 header 버튼 처리 4. map화면 우측 층 페이징처리 및 정상동작하게 수정

+++ src/main/java/com/pluxity/ktds/domains/building/service/BuildingService.java
@@ -85,7 +85,8 @@ public class BuildingService {
     @Transactional(readOnly = true)
     public BuildingDetailResponseDTO findOutdoorDetail() {
 
-        Building building = buildingRepository.findTop1ByIsIndoorOrderByIdDesc("N");
+        Building building = buildingRepository.findTop1ByIsIndoorOrderByIdDesc("N")
+                .orElseThrow(() -> new CustomException(NOT_FOUND_BUILDING));
 
         return building.toDetailResponseDTO();
     }
@@ -168,11 +169,14 @@ public class BuildingService {
         buildingFileHistoryRepository.deleteAll(buildingFileHistories);
 
         List<Patrol> patrols = patrolRepository.findByBuildingId(building.getId());
-        for (Patrol patrol : patrols) {
-            List<PatrolPoint> patrolPoints = patrolPointRepository.findByPatrolId(patrol.getId());
-            patrolPointRepository.deleteAll(patrolPoints);
+
+        if (!patrols.isEmpty()) {
+            for (Patrol patrol : patrols) {
+                List<PatrolPoint> patrolPoints = patrolPointRepository.findByPatrolId(patrol.getId());
+                patrolPointRepository.deleteAll(patrolPoints);
+            }
+            patrolRepository.deleteByBuildingId(building.getId());
         }
-        patrolRepository.deleteByBuildingId(building.getId());
 
         buildingRepository.delete(building);
     }
